backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.2
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
python-decouple==3.8
openpyxl==3.1.5
reportlab==4.2.2
Pillow==10.4.0
gunicorn==22.0.0
ruff==0.5.5
black==24.8.0
isort==5.13.2
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.1
